requirements.txt
sklearn
nltk
PyEnchant
python-weka-wrapper3
numpy
scipy
gensim
nltk
beautifulsoup4
gevent
newspaper3k
future
pymysql
pandas
GoogleScraper
google
flask
flask_cors
gunicorn
pyvirtualdisplay
line-bot-sdk